requirements.txt
acme==1.1.0
alembic==1.9.1
authentication==1.1.0
autopep8==2.0.1
Django==3.2.9
django-ckeditor==6.2.0
django-cors-headers==3.10.0
django-crispy-forms==1.11.1
django-crontab==0.7.1
django-datetimepicker==3.14
django-email-users==1.1.0
django-extensions==3.1.5
django-filter==21.1
django-import-export==2.7.1
django-indian-numbers==0.0.2
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-taggit==1.4.0
django-taggit-autosuggest==0.3.8
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
greenlet==2.0.1
gunicorn==20.1.0
html2excel==0.0.6
html5lib==1.1
httplib2==0.14.0
idna==3.3
pytz==2021.3
pyxdg==0.26
soupsieve==2.2.1
SQLAlchemy==1.4.45
sqlparse==0.4.2
tlslite==0.4.9

